Update files with abstract method
